// frontend/src/openapi/models/EvalSampleIO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InputSourceEnum } from './InputSourceEnum';
import {
    InputSourceEnumFromJSON,
    InputSourceEnumFromJSONTyped,
    InputSourceEnumToJSON,
    InputSourceEnumToJSONTyped,
} from './InputSourceEnum';

/**
 *
 * @export
 * @interface EvalSampleIO
 */
export interface EvalSampleIO {
    /**
     *
     * @type {any}
     * @memberof EvalSampleIO
     */
    input: any | null;
    /**
     *
     * @type {InputSourceEnum}
     * @memberof EvalSampleIO
     */
    inputSource: InputSourceEnum;
    /**
     *
     * @type {any}
     * @memberof EvalSampleIO
     */
    output: any | null;
    /**
     *
     * @type {Array<{ [key: string]: any; }>}
     * @memberof EvalSampleIO
     */
    outputMessages: Array<{ [key: string]: any; }>;
    /**
     *
     * @type {any}
     * @memberof EvalSampleIO
     */
    reference: any | null;
    /**
     *
     * @type {boolean}
     * @memberof EvalSampleIO
     */
    truncated: boolean;
}



/**
 * Check if a given object implements the EvalSampleIO interface.
 */
export function instanceOfEvalSampleIO(value: object): value is EvalSampleIO {
    if (!('input' in value) || value['input'] === undefined) return false;
    if (!('inputSource' in value) || value['inputSource'] === undefined) return false;
    if (!('output' in value) || value['output'] === undefined) return false;
    if (!('outputMessages' in value) || value['outputMessages'] === undefined) return false;
    if (!('reference' in value) || value['reference'] === undefined) return false;
    if (!('truncated' in value) || value['truncated'] === undefined) return false;
    return true;
}

export function EvalSampleIOFromJSON(json: any): EvalSampleIO {
    return EvalSampleIOFromJSONTyped(json, false);
}

export function EvalSampleIOFromJSONTyped(json: any, ignoreDiscriminator: boolean): EvalSampleIO {
    if (json == null) {
        return json;
    }
    return {

        'input': json['input'],
        'inputSource': InputSourceEnumFromJSON(json['input_source']),
        'output': json['output'],
        'outputMessages': json['output_messages'],
        'reference': json['reference'],
        'truncated': json['truncated'],
    };
}

export function EvalSampleIOToJSON(json: any): EvalSampleIO {
    return EvalSampleIOToJSONTyped(json, false);
}

export function EvalSampleIOToJSONTyped(value?: EvalSampleIO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'input': value['input'],
        'input_source': InputSourceEnumToJSON(value['inputSource']),
        'output': value['output'],
        'output_messages': value['outputMessages'],
        'reference': value['reference'],
        'truncated': value['truncated'],
    };
}